many updates CSS and LESS files and generally a responsive scaling overhaul of the settings page

+++ lesstocss.sh
@@ -3,7 +3,10 @@
 echo 'converting LESS to CSS...'
 # generate CSS from LESS
 lessc less/styles.less css/styles.css
-rm css/styles.min.css
+
+if [ -a css/styles.min.css ]; then
+  rm css/styles.min.css
+fi
 
 if [ -n "$1" ]; then
   if [ $1 = '--prefix' ] || [ $1 = '-p' ]; then
@@ -12,14 +15,21 @@ if [ -n "$1" ]; then
     # prefixed (-s overwrites existing file)
     prefixr --input ./css/styles.css -s
   fi
-    if [ $1 = '--minify' ] || [ $1 = '-m' ]; then
-    echo 'creating minified and prefixed...'
+  if [ $1 = '--minify' ] || [ $1 = '-m' ]; then
 
-    # copy to .min.css to create a minified version as well
-    cp css/styles.css css/styles.min.css
+    if [ ! -n "$2" ]; then
+      echo 'creating minified....'
+      lessc --yui-compress less/styles.less css/styles.min.css
+    else
+      echo 'creating minified and prefixed...'
 
-    # prefixed and minified (-s overwrites existing file)
-    prefixr --input ./css/styles.min.css -s -c
+      # copy to .min.css to create a minified version as well
+      cp css/styles.css css/styles.min.css
+
+      # prefixed and minified (-s overwrites existing file)
+      prefixr --input ./css/styles.min.css -s -c
+   fi
+
   fi
 fi
